Reorganized files
Moved Activities into view directory and moved TextWatchers into view/validator directory
