+++ app/src/main/java/com/group1/project3/view/EditCardFragment.java
@@ -1,17 +1,17 @@
-package com.group1.project3;
+package com.group1.project3.view;
 
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.RecyclerView;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.EditText;
 import android.widget.TextView;
 
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.google.gson.Gson;
+import com.group1.project3.R;
 import com.group1.project3.model.Card;
 
 public class EditCardFragment extends Fragment {
