+++ app/src/main/java/com/group1/project3/view/MainActivity.java
@@ -1,4 +1,4 @@
-package com.group1.project3;
+package com.group1.project3.view;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,6 +6,7 @@ import android.view.View;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.group1.project3.R;
 import com.group1.project3.util.FirebaseUtil;
 
 /**
