+++ app/src/main/java/com/group1/project3/view/PreviewCardFragment.java
@@ -1,15 +1,15 @@
-package com.group1.project3;
+package com.group1.project3.view;
 
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import androidx.fragment.app.Fragment;
+
 import com.google.gson.Gson;
+import com.group1.project3.R;
 import com.group1.project3.model.Card;
 
 public class PreviewCardFragment extends Fragment {
