+++ app/src/main/res/layout/activity_create_project.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    tools:context=".CreateProjectActivity">
+    tools:context=".view.CreateProjectActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
