+++ app/src/main/res/layout/activity_profile.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="@dimen/main_padding"
-    tools:context=".ProfileActivity">
+    tools:context=".view.ProfileActivity">
 
     <LinearLayout
         android:id="@+id/linearLayout"
