+++ app/src/main/res/layout/activity_register.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="@dimen/main_padding"
-    tools:context=".RegisterActivity">
+    tools:context=".view.RegisterActivity">
 
     <com.google.android.material.textfield.TextInputLayout
         android:layout_width="match_parent"
