+++ app/src/main/res/layout/fragment_preview_card.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PreviewCardFragment">
+    tools:context=".view.PreviewCardFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
